refactor: rearrange files to use single css for all buttons
