--- src/components/view/categories/categories.css
@@ -1,51 +1,0 @@
-.categories {
-    display: flex;
-    flex-wrap: nowrap;
-    width: 100%;
-    height: 80px;
-    overflow: auto;
-    align-items: center;
-    font: 300 1em 'Fira Sans', sans-serif;
-    -ms-overflow-style: none;
-    scrollbar-width: none;
-}
-.categories::-webkit-scrollbar {
-    display: none;
-}
-
-.category__item {
-    background: none;
-    border: 2px solid #30c5ff;
-    border-radius: 10px;
-    font: inherit;
-    line-height: 1;
-    margin: 0.5em;
-    padding: 1em 2em;
-    color: #70d6ff;
-    transition: 0.25s;
-    cursor: pointer;
-}
-
-.category__item:hover,
-.category__item:focus {
-    border-color: #3fcc59;
-    color: #69db7e;
-    box-shadow: 0 0.5em 0.5em -0.4em #3fcc59;
-    transform: translateY(-0.25em);
-}
-
-.category__item-name {
-    font-weight: 400;
-    white-space: nowrap;
-}
-
-@media screen and (max-width: 500px) {
-    .categories{
-        height: 60px;
-    }
-    .category__item {
-        font-size: 0.8em;
-        border-radius: 5px;
-        padding: 1em 1.2em;
-    }
-}
